+++ cluster/cluster/failover_cluster_invoker.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package cluster_impl
+package cluster
 
 import (
 	"strconv"
