+++ cluster/cluster/failover_cluster_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package cluster_impl
+package cluster
 
 import (
 	"context"
