+++ config_center/apollo/mockDubbog.properties.json
@@ -1,0 +1,1 @@
+{"appId":"testApplication_yang","cluster":"default","namespaceName":"mockDubbog.properties","releaseKey":"20191104105242-0f13805d89f834a4","configurations":{"registries.hangzhouzk.username":"11111"}}
